Move config reading to settings_window.py
